Make MatchDescription its own comp
