+++ src/components/MatchDescription.tsx
@@ -1,0 +1,33 @@
+import { notEmpty } from "../lib";
+import { type Match } from "../lib/rating";
+import { MatchDetails } from "../pages/admin/MatchDetails";
+import { TeamDetails } from "../pages/admin/TeamDetails";
+
+export const MatchDescription = ({ match }: { match: Match | undefined }) => {
+  if (match === undefined) {
+    return <></>;
+  }
+
+  const teamPlayers = {
+    black: [match.blackPlayerOne.name, match.blackPlayerTwo?.name].filter(
+      notEmpty,
+    ),
+    white: [match.whitePlayerOne.name, match.whitePlayerTwo?.name].filter(
+      notEmpty,
+    ),
+  };
+
+  return (
+    <>
+      <div class="flex flex-col justify-between gap-3 lg:flex-row">
+        <TeamDetails title="Team White" team={teamPlayers.white} />
+        <TeamDetails title="Team Black" team={teamPlayers.black} />
+      </div>
+      <MatchDetails
+        result={match.result}
+        scoreDiff={match.scoreDiff}
+        dateLogged={match.createdAt}
+      />
+    </>
+  );
+};
